sql_example_endpoints/requirements.txt
uvicorn
fastapi
psycopg2-binary
python-dotenv